+++ test/index/fix_links_with_anchors_test.js
@@ -12,8 +12,12 @@ describe('index/fix links with anchors:', function () {
     this.files = {
       'docs/README.md': {
         contents:
+          '* [Hi](../README.md)\n' +
           '* [Introduction](/docs/introduction.md)\n' +
           '* [Getting started](/docs/getting-started.md)\n'
+      },
+      'README.md': {
+        contents: '# hi'
       },
       'docs/getting-started.md': {
         contents: 'hi'